K-fold Cross-validation for Cooperative AFT Model Estimation

Function cv.coop performs k-fold cross-validation for cooperative estimation in Accelerated Failure Time (AFT) models. This function helps selecting the optimal tuning parameters (lambda) by minimizing prediction error across folds. It is typically used by function aft_coop to select the optimal parameter.

cv.coop(
  X,
  Xtilde,
  Y,
  delta,
  sigma = NULL,
  lambda = NULL,
  rho = NULL,
  nfolds = 5,
  model,
  parallel = TRUE,
  ncore_max = 5,
  seed = 123
)

Arguments

X

Matrix of combined covariates of dimension n x p_u+p_z obtained as X=[U,Z], when two views are provided (case=="coop"). Otherwise X=[U] or X=[Z], when case=="onlyU" or case=="onlyZ", respectively.

Xtilde

Matrix of cooperative adjusted covariates for cooperative estimation of dimension n x p_u+p_z obtained as Xtilde=[U,-Z], when two views are provided (case=="coop");

Y

Vector of log observed survival times or log censored times (response variable in log scale) of dimension n.

delta

Vector of censoring indicators (1 for event, 0 for censored) of dimension n.

sigma

Scale parameter for the AFT model.

lambda

Vector of regularization parameters in descending order.

rho

scalar cooperation parameter or vector of cooperation parameters, the latter option works when case=="coop".

nfolds

Number of cross-validation folds (default: 5)

model

Character string specifying the AFT model type. The following choices are available: "weibull";"lognormal";"loglogistic".

parallel

Logical; whether to use parallel processing (default: TRUE)

ncore_max

Maximum number of cores for parallel processing (default: 5)

seed

Random seed for reproducibility (default: 123)

Value

A list of class "cv.coop" containing:

cv.err.linPred: Vector of cross-validation errors for each lambda

cv.err.obj: Matrix of fold-specific errors (rows are folds, columns are lambdas)

lambda_grid: vector of values used in the grid of lambda for the cross validation

lambda.min: Value of lambda that minimizes cross-validation error

ind.lambda.min: Index of lambda.min in the lambda vector

lambda.1se: Largest lambda within one standard error of minimum

ind.lambda.1se: Index of lambda.1se in the lambda vector

case.rho: chosed value of parameter rho

Details

The function implements k-fold cross-validation by:

Randomly partitioning the data into k folds

For each fold:

Fitting the model on training data (excluding the fold)

Computing predictions on validation data (the fold)

Calculating prediction error

Computing both minimum error lambda (lambda.min) and 1-standard-error rule lambda (lambda.1se)

The function requires the cvTools package for fold generation. The number of cores used for parallel processing is automatically determined based on system availability, number of folds and user-specified maximum. Parallel processing implementation differs between Windows (using makeCluster) and Unix-like systems (using mclapply).

Note

Last change 25/03/2025

See also