Variable screening based on ranking by marginal utility

This function performs marginal ranking of features based on their association with the survival outcome (see, for example, Fan and Lv (2008) and Fan, Feng and Wu (2010)). Specifically, it computes the marginal coefficient for each feature in the dataset and s orts the covariates in decreasing order according to the selected ranking method: absmg, mg, or mgpadj.

marginal_ranking(
  X,
  y,
  delta,
  family = c("cox", "AFT"),
  model = c("weibull", "exponential", "lognormal", "loglogistic"),
  rank = c("absmg", "mg", "mgpadj")
)

Arguments

X

Matrix of covariates correspondig to the view of dimension n x p_{view}, where n is the number of samples and p_{view} is the number of variables in the view.

y

Vector of observed survival times in the time-domain or censored times of dimension n, where n is the number of samples.

delta

Vector of censoring indicators (1 for event, 0 for censored) of dimension n, where n is the number of samples.

family

Character string indicating the type of survival model to fit. Options are cox for Cox Proportional Hazards model and AFT for Accelerated Failure Time model.

model

Character string specifying the specific AFT model to use. It can assume one of the following values: "weibull";"lognormal";"loglogistic". Ignored if family = "cox".

rank

Character string specifying the ranking method. It can assume one of the following values: "absmg", "mg" and "mgpadj".

Value

A data frame with the ranked features. The columns include:

ranking: The rank of the feature based on the selected method.

symbol: The name of the feature.

rank: The ranked marginal utility values using one of several methods: absmg, mg, mgpadj.

Details

This function implements marginal ranking for the features of a matrix X based on their association with survival outcomes in y,delta , using either the Cox Proportional Hazards or Accelerated Failure Time (AFT) models. The functions assume that the feature in X were already preprocessed to remove irrelevant information and normalized, if necessary. Such preprocessing depends on the type of variables encoded in matrix X. Then, it uses the margcoef function to compute the marginal coefficients for each feature, employing several ranking methods (namely,absmg, mg, or mgpadj) based on marginal utility or adjusted p-values. The margcoef function, in turn, uses the mg function to calculate the regression coefficients and adjusted p-values, depending on the specified family (Cox or AFT). The ranking of the features can be done according to one of the following criteria:

absmg: The absolute marginal utility function.

mg: The marginal utility function.

mgpadj: The p-value adjusted according with the marginal utility function.

When choosing mgpadj, the adjustment method includes the Benjamini and Hochberg (1995) (BH or its alias fdr) correction.

Note

Last change 08/04/2025

References

Benjamini, Y., and Hochberg, Y. (1995). "Controlling the false discovery rate: a practical and powerful approach to multiple testing", Journal of the Royal Statistical Society Series B, 57, 289–300. DOI: 10.1111/j.2517-6161.1995.tb02031.x

Fan, J., Feng, Y., and Wu, Y. (2010). "High-dimensional Variable Selection for Cox Proportional Hazards Model". IMS Collections, 6, 70-86. DOI: 10.1214/10-IMSCOLL606

Fan, J., and Lv, J. (2008). "Sure independence screening for ultrahigh dimensional feature space", Journal of the Royal Statistical Society Series B: Statistical Methodology, 70(5), 849-911. DOI: 10.1111/j.1467-9868.2008.00674.x

Examples

if (FALSE) { # \dontrun{
# Generate example data
set.seed(123)
model <- "weibull"
data <- generate_data(
  model = model,
    n = 200,
    pu = 150,
    pz = 150,
    tu = 6,
    tz = 6,
    rate = 40,
    sigma_true = 0.5
    )

y <- data$times_c    # vector of survival times or censored times
delta <- data$delta  # vector of censoring indicators
U <- data$U  # matrix corresponding to the first view
colnames(U) <- paste0("feature_",  1:ncol(U))
Z <- data$Z  # matrix corresponding to the second view
colnames(Z) <- paste0("feature_",  1:ncol(Z))

screening_U <- marginal_ranking(U, y, delta, family = "AFT", model = model, rank = "absmg")
print(screening_U)
screening_Z <- marginal_ranking(U, y, delta, family = "AFT", model = model, rank = "absmg")
print(screening_U)
} # }