Plot AFT Cooperative Cross-Validation Results

Internal function that creates a ggplot visualization of cross-validation results from an accelerated failure time (AFT) cooperative regularization model.

plot_aft_coop.cv(cv.out, nfolds, case, addbar = F)

Arguments

cv.out

List containing cross-validation results obtained by cv.coop , must include components:

cv.err.obj - Matrix of CV errors for each fold and lambda value

cv.err.linPred - Vector of mean CV errors for linear predictor

lambda_grid Numeric vector of lambda values used in the regularization path.

lambda.1se - Lambda value within 1 standard error of minimum CV error

lambda.min - Lambda value with minimum CV error

case.rho - Rho value for the current case

nfolds

Integer specifying the number of cross-validation folds used.

case

Character or numeric identifier for the current case being analyzed.

addbar

Logical to add error bar to the crossvalidation curve (Default FALSE)

Value

A ggplot object visualizing cross-validation results

Details

The function calculates standard errors for cross-validation error estimates and creates a visualization with error bars. It marks the lambda.min and lambda.1se values with vertical dashed lines. The object cv.err.obj is obtained using cv.coop

Note

Last change 15/04/2025