Evaluation of the linear predictor score for the atf_coop Survival Model

This function computes the linear predictor for the cooperative AFT survival model, where the linear predictor is defined as \(X \hat{\beta}\) with \(X\) being the regression matrix and \(\hat{\beta}\) the vector of estimated parameters obtained from the aft_coop function. The linear predictor score can be used as Prognostic index to stratify patients in risk classes or to compute C-Index.

predict_aft_coop(
  U,
  Z,
  mU = NULL,
  mZ = NULL,
  beta,
  case = c("onlyU", "onlyZ", "coop")
)

Arguments

U

A matrix of feature variables (e.g., gene expression data). Each row corresponds to a sample, and each column corresponds to a feature (gene). It should have dimensions n x pu.

Z

A matrix of additional feature variables (e.g., methylation data). Each row corresponds to a sample, and each column corresponds to a feature (attribute). It should have dimensions n x pz.

mU

a vector of parameters for centering the U matrix. (default mU=colMeans(U), but when predicting on the test set the value mus be adjusted to those of the training set.)

mZ

a vector of parameters for centering the Z matrix. (default mZ=colMeans(Z),but when predicting on the test set the value mus be adjusted to those of the training set.)

beta

A vector of estimated coefficients (of length (pu), (pz) or (pu + pz), depending on the case) obtained from obtained from the aft_coop function.

case

A character string specifying the case for prediction. Options are:

"onlyU"Predict using only the U matrix.

"onlyZ"Predict using only the Z matrix.

"coop"Predict using both the U and Z matrices combined.

Value

A vector of predicted linear scores.

Details

The depending on the estimation strategy of the regression coefficients, the prediction can be made based on:

"onlyU": When only the U matrix (e.g., gene expression data) was used to estimate \(\hat{\beta}\).

"onlyZ": When only the Z matrix (e.g., methylation data) was used to estimate \(\hat{\beta}\).

"coop": When the two matrices U and Z matrices were used to estimate \(\hat{\beta}\).

Moreover, when applied to the test set the regression matrices are centered according to the mean of the training set. This is achieved using the parameter mU and mZ.

Note

Last change 11/03/2025

Examples