Variable Screening method

This function selects a subset of relevant features from a set of potential features based on their association with patient survival, thus allowing a preliminary reduction in the number of variables before applying a penalized regression approach. The function uses the following ranking methods:

absmg: The marginal utilities ordered from largest to smallest.

mg: The regression coefficients ordered from largest to smallest, with the top and bottom perc selected.

mgpadj: The marginal utilities with p-values adjusted to be less than thresh.

variable_screening(
  X,
  y,
  delta,
  family = c("cox", "AFT"),
  model = c("weibull", "loglogistic", "lognormal"),
  rank = c("absmg", "mg", "mgpadj"),
  topn = 500,
  perc = 5,
  thresh = 0.5
)

Arguments

X

Matrix of covariates correspondig to the view of dimension n x p_{view}, where n is the number of samples and p_{view} is the number of variables in the view.

y

Vector of observed survival times or censored times of dimension n, where n is the number of samples.

delta

Vector of censoring indicators (1 for event, 0 for censored) of dimension n, where n is the number of samples.

family

Character string indicating the type of survival model to fit. Options are cox for Cox Proportional Hazards model and AFT for Accelerated Failure Time model.

model

Character string specifying the specific AFT model to use. It can assume one of the following values "weibull";"lognormal";"loglogistic". Ignored if family = "cox".

rank

Character string specifying the ranking method. It can assume one of the following values:"absmg", "mg" or "mgpadj"

topn

An integer specifying the number of top features to return (default 500). Used when rank = "absmg".

perc

An integer specifying the percentage of top and bottom features to return (default 5). Used when rank = "mg".

thresh

A numeric value representing the threshold for adjusted p-values. Only features with adjusted p-values smaller than thresh will be considered. Used when rank = "mgpadj".

Value

A vector of indices corresponding to the topn ranked features when rank = "absmg", the top and bottom features when rank = "mg", and the features with adjusted p-values smaller than thresh when rank = "mgpadj".

Details

The function performs variable screening based on the marginal ranking of the features in matrix X, thus reducing the features' dimensionality while retaining relevant ones. :

absmg: It uses the absolute marginal utility function and selects the topn features as screened variables.

mg: It uses the marginal utility function and selects as screened variables given percentage perc of the top or bottom features.

mgpadj: It uses the p-value adjusted according to the marginal utility function to select as screened variables those with adjusted p-values less than thresh.

The adjustment method includes the Benjamini and Hochberg (1995) (BH or its alias fdr) correction.

Note

Last change 08/04/2025

See also

Examples

if (FALSE) { # \dontrun{
# Generate example data
set.seed(123)
model <- "weibull"
topn <- 500
data <- generate_data(
  model = model,
    n = 200,
    pu = 150,
    pz = 150,
    tu = 6,
    tz = 6,
    rate = 40,
    sigma_true = 0.5
    )

y <- data$times_c    # vector of survival times or censored times
delta <- data$delta  # vector of censoring indicators
U <- data$U  # matrix corresponding to the first view
colnames(U) <- paste0("feature_",  1:ncol(U))
Z <- data$Z  # matrix corresponding to the second view
colnames(Z) <- paste0("feature_",  1:ncol(Z))

index_screen_U <- variable_screening(U, y, delta, family = "AFT", model = model, rank = "absmg", topn = topn)
index_screen_Z <- variable_screening(Z, y, delta, family = "AFT", model = model, rank = "absmg", topn = topn)
} # }